b64encode.c, hexify.c: clarify casts and integer promotion
- b64encode.c: clang happily determines that (size_t)&0x3f can be safely
  cast to (uint32_t), but gcc wants this to be explicit.

- hexify.c: oh boy.  Fun fact about C99: due to integer promotion, you
  cannot use compound assignment for values smaller than an int without
  implicitly casting the larger (int) type into the smaller type.

  Concretely, given
      uint8_t a;
      a += 1;
  we're implicitly doing:
      a = (int)a + 1;

  In the case of unhexify(), the previous code used <<= and +=.  This
  commit changes those to simple assignment, with explicit casts as
  necessary.

+++ libcperciva/util/b64encode.c
@@ -94,7 +94,7 @@ b64decode(const char * in, size_t inlen, uint8_t * out, size_t * outlen)
 		for (t = 0, i = 0; i < 4; i++) {
 			t <<= 6;
 			pos = strchr(b64chars, in[i]) - b64chars;
-			t += pos & 0x3f;
+			t += (uint32_t)(pos & 0x3f);
 		}
 
 		/* Output 3 bytes. */
